move methods into wang.mjs for reuse
